+++ src/components/layout/Navbar.tsx
@@ -7,7 +7,7 @@ import { GITHUB_URL } from '@/domain/owner/constant'
 import useScrollDirection from '@/hooks/useScrollDirection'
 import { Z_INDEX } from '@/styles/constant'
 import { convertHEXToRGB } from '@/utils/string'
-import { PAGE_PATH } from '@/constants/common'
+import { PAGE_PATH } from '@/constants/route'
 import useScrollTop from '@/hooks/useScrollTop'
 
 import Frame from './Frame'
