+++ src/constants/layout.ts
@@ -1,0 +1,5 @@
+export const DEVICE_SIZE = {
+  mobile: 375,
+  tablet: 768,
+  pc: 1024,
+} as const
