+++ src/constants/route.ts
@@ -1,9 +1,3 @@
-export const DEVICE_SIZE = {
-  mobile: 375,
-  tablet: 768,
-  pc: 1024,
-} as const
-
 export const PAGE_PATH = {
   About: {
     label: 'About',
